let the user specify a test that make will run
Reviewed-on: http://review.couchbase.org/9534
Tested-by: Farshid Ghods <farshid.ghods@gmail.com>

+++ scripts/start_cluster_and_run_tests.sh
@@ -25,7 +25,12 @@ cd ../ns_server
 ./cluster_run --nodes=$servers_count &> $wd/cluster_run.log &
 pid=$!
 popd
-./testrunner -c $conf -i $ini 2>&1 | tee make_test.log
+
+if [[ -f $conf ]] ; then
+    ./testrunner -c $conf -i $ini 2>&1 | tee make_test.log
+else
+    ./testrunner -t $conf -i $ini 2>&1 | tee make_test.log
+fi
 kill $pid
 wait
 ! grep FAILED make_test.log &> /dev/null
